Servis povezan sa RabbitMQ brokerom(zapoceto samo), salje asinhrone poruke servisu za notifikacije

+++ src/main/java/raf/sk/userservice/dto/rank/RankResponseDto.java
@@ -10,5 +10,5 @@ import lombok.Setter;
 public class RankResponseDto {
 
     private String type;
-    private Integer discount;
+    private Double discount;
 }
